+++ src/NodeVisitor/TestCaseGeneratorNodeVisitor.php
@@ -34,6 +34,7 @@ use RegexParser\Node\LimitMatchNode;
 use RegexParser\Node\LiteralNode;
 use RegexParser\Node\PcreVerbNode;
 use RegexParser\Node\PosixClassNode;
+use RegexParser\Node\QuantifierBounds;
 use RegexParser\Node\QuantifierNode;
 use RegexParser\Node\RangeNode;
 use RegexParser\Node\RegexNode;
@@ -499,19 +500,9 @@ final class TestCaseGeneratorNodeVisitor extends AbstractNodeVisitor
      */
     private function parseQuantifierRange(string $q): array
     {
-        return match ($q) {
-            '*' => [0, null],
-            '+' => [1, null],
-            '?' => [0, 1],
-            default => preg_match('/^\{(\d++)(?:,(\d*+))?\}$/', $q, $m) ?
-                (isset($m[2]) ? ('' === $m[2] ?
-                    [(int) $m[1], null] :
-                    [(int) $m[1], (int) $m[2]]
-                ) :
-                    [(int) $m[1], (int) $m[1]]
-                ) :
-                [1, 1],
-        };
+        $bounds = QuantifierBounds::parse($q);
+
+        return null === $bounds ? [1, 1] : [$bounds->min, $bounds->max];
     }
 
     /**